Add stuff to build/run device properties binary
This is just a Makefile to build the code given by the professor
and a `sbatch` script to execute tha binary inside a node and get
the GPU information.
In addition it includes a minor fix for some offeset
arguments in a printing line that does not allow to build
the program correctly, once the printing section was
uncommented.

Signed-off-by: mario carrillo <carrillo@rhrk.uni-kl.de>