Fix exercise 04 implementation
The exercise 04 was wrong implemented. The commit removes the
the interation over structures and replace them by linear arrays.
In addition, the blocks and threads are computed according to
the number of particles. Finally, it removes the unecessary
while loop and the iterations over off incerment. The execution
time was reduced from 771 seconds to 175 seconds approximatly.

Signed-off-by: mario carrillo <carrillo@rhrk.uni-kl.de>